requirements.txt
diffusers==0.33.1
numpy==2.1.2
opencv-python==4.11.0.86
pillow==11.0.0
segment-anything==1.0
torch==2.5.1+cu121
torchvision==0.20.1+cu121
xformers==0.0.29.post1 